move connectDB to ./utils/connectDB
